requirements.txt
mysql-connector==2.2.9
python-telegram-bot==12.0.0b1
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.3
uuid==1.30
dateparser
sqlalchemy
